Corrige de bugs
Bugs corrigidos:
- Ao selecionar um pixel muito claro, ele + 5% causava um numero maior
que 255, causando overflow de uchar.
- Pressionar uma tecla nao fecha a aplicacao.

+++ include/Image.hpp
@@ -19,7 +19,7 @@ class Image {
     Image ();
     cv::Mat image;
     cv::Mat paintedImage;
-    uchar marginColor[2][3] = {0};
+    int marginColor[2][3] = {0};
     bool isGrayscale;
 };
